% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppMP.R
\docType{package}
\name{RcppMP-package}
\alias{RcppMP-package}
\title{R multi-precision types for Rcpp.}
\description{
R multi-precision types for Rcpp.
}
\references{
Extending Rcpp:
 \url{http://cran.r-project.org/web/packages/Rcpp/vignettes/Rcpp-extending.pdf}

The GNU Multiple Precision Arithmetic Library
 \url{https://gmplib.org/}

The GNU MPFR Library
 \url{http://www.mpfr.org/}

MPFR C++ mpreal
 \url{http://www.holoborodko.com/pavel/mpfr/}

Boost.Multiprecision
 \url{http://www.boost.org/doc/libs/release/libs/multiprecision/doc/html/index.html}
}
\seealso{
\link[Rcpp]{Rcpp} \link[gmp:bigz]{gmp bigz}
 \link[gmp:bigq]{gmp bigq} \link[Rmpfr]{Rmpfr} \link[BH]{BH}
}
\keyword{Rcpp}
\keyword{multi-precision}

