% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcppmpVersions}
\alias{rcppmpVersions}
\title{Library versions.}
\usage{
rcppmpVersions()
}
\value{
List of named character vectors indicating the current
  versions.
}
\description{
Available multi-precision library versions.
}
