% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_cpp}
\alias{test_cpp}
\title{Console access to c++ tests.}
\usage{
test_cpp(args = "")
}
\arguments{
\item{args}{Single or multiple character strings.}
}
\description{
Console access to c++ tests.
}

